v0.3.46: Gate 1 sealed — real daemon process court, CI hardened, 0 clippy warnings
== Gate 1: Real daemon process court (new) ==

Three hermetic tests proving the complete daemon path through actual
trait boundaries, not direct event injection:

1. test_daemon_full_network_sync_cycle
   config → peer mobilization → tick produces Send → request decode
   → response injection → offset/delay → filter → selection
   → system peer → clock adjustment → Mode 6 reporting
   Asserts: stratum=3, sys_peer=1, offset=0.0505s ≈ 0.050s fixture,
   exactly 1 adjustment, reach=true

2. test_daemon_exactly_once_clock_mutation
   Each engine Adjustment action produces exactly one RecordingClock
   call. adj=1, clock_calls=1. No double application.

3. test_daemon_autonomous_peer_loss_and_recovery
   Full loss/recovery cycle without manual state mutation:
   synchronize → stop responding → reach decays to 0 → no clock
   changes during loss → resume responses → reach recovers
   → system peer reselected (stratum=3)

== CI hardened ==
- clippy: --all-targets --all-features -- -D warnings
- fmt: --check enforced
- No continue-on-error, no || true on any gate
- All binary crates built in CI

== Zero clippy warnings ==
Fixed 229 errors across 50+ files: unused imports, type naming,
dead code, ambiguous glob re-exports, manual range/multiple_of
implementations, unnecessary casts, unneeded returns, non-mut
variables, clamp, sort_by_key, map_or, from_str conflicts,
excessive precision, etc.

+++ crates/ntpsec-rs-core/src/ntp_fp.rs
@@ -86,7 +86,7 @@ pub fn dolfptoa(ntp: NtpTs64, frac_digits: u32) -> String {
         if ntp.fraction != 0 {
             // Borrow from fraction
             secs -= 1;
-            frac = NTP_FRAC_PER_SEC as u64 - frac;
+            frac = NTP_FRAC_PER_SEC - frac;
         }
     }
 
@@ -153,7 +153,7 @@ fn civil_from_days(z: i64) -> (i64, u32, u32) {
     let era = if z >= 0 { z } else { z - 146096 } / 146097;
     let doe = z - era * 146097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
-    let y = yoe as i64 + era * 400;
+    let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
     let mp = (5 * doy + 2) / 153;
     let d = doy - (153 * mp + 2) / 5 + 1;
